add login and register page
