+++ frontend/src/src/views/authentication/Login.js
@@ -2,9 +2,9 @@ import React from 'react';
 import {Link as RouterLink} from 'react-router-dom';
 import {Card, CardContent, Divider, Grid, Link, makeStyles, Typography, useMediaQuery, useTheme} from '@material-ui/core';
 
-import FirebaseRgister from './../firebase-forms/FirebaseRgister';
+import LoginFrom from './login/LoginForm';
 
-import logo from './../../../../assets/images/logo.svg';
+import logo from './../../assets/images/logo.svg';
 
 const useStyles = makeStyles((theme) => ({
     root: {
@@ -43,13 +43,13 @@ const useStyles = makeStyles((theme) => ({
     }
 }));
 
-const Register = () => {
+const Login = () => {
     const classes = useStyles();
     const theme = useTheme();
     const matchDownSM = useMediaQuery(theme.breakpoints.down('sm'));
 
     return (
-        <Grid container justifyContent={matchDownSM ? 'center' : 'space-between'} alignItems="flex-start" className={classes.root}>
+        <Grid container justifyContent={matchDownSM ? 'center' : 'space-between'} alignItems="center" className={classes.root}>
             <Grid item xs={12} sx={{minHeight: '100vh', height: '100%'}}>
                 <Grid
                     sx={{minHeight: '100vh', height: '100%', p: matchDownSM ? 0 : '0 80px'}}
@@ -84,7 +84,7 @@ const Register = () => {
                                                                 gutterBottom
                                                                 variant={matchDownSM ? 'h3' : 'h2'}
                                                             >
-                                                                Sign up
+                                                                Hi, Welcome Back
                                                             </Typography>
                                                         </Grid>
                                                     </Grid>
@@ -99,7 +99,7 @@ const Register = () => {
                                         </Grid>
                                     </Grid>
                                     <Grid item xs={12}>
-                                        <FirebaseRgister />
+                                        <LoginFrom login={3} />
                                     </Grid>
                                     <Grid item xs={12}>
                                         <Divider />
@@ -108,11 +108,11 @@ const Register = () => {
                                         <Grid item container direction="column" alignItems="center" xs={12}>
                                             <Typography
                                                 component={RouterLink}
-                                                to="/pages/login/login3"
+                                                to="/register"
                                                 variant="subtitle1"
                                                 className={classes.title}
                                             >
-                                                Having an account?
+                                                Don't have an account?
                                             </Typography>
                                         </Grid>
                                     </Grid>
@@ -129,7 +129,7 @@ const Register = () => {
                         alignItems="center"
                         spacing={matchDownSM ? 2 : 0}
                     >
-                       <Grid item>
+                        <Grid item>
                             <Typography component={Link} href='https://berrydashboard.io' target='_blanks' variant="subtitle1" color={theme.palette.grey[600]}>
                                 www.berrydashboard.io
                             </Typography>
@@ -146,4 +146,4 @@ const Register = () => {
     );
 };
 
-export default Register;
+export default Login;
