+++ frontend/src/src/views/authentication/login/LoginForm.js
@@ -1,8 +1,8 @@
-import React, {useEffect} from 'react';
-import {useSelector} from 'react-redux';
+import React from 'react';
 import clsx from 'clsx';
 import * as Yup from 'yup';
 import {Formik} from 'formik';
+import {useSelector} from 'react-redux';
 import {
     Box,
     Button,
@@ -17,18 +17,13 @@ import {
     InputLabel,
     makeStyles,
     OutlinedInput,
-    TextField,
     Typography
 } from '@material-ui/core';
 import {Link} from 'react-router-dom';
 
-import useScriptRef from '../../../../hooks/useScriptRef';
+import useScriptRef from '../../../hooks/useScriptRef';
 import Visibility from '@material-ui/icons/Visibility';
 import VisibilityOff from '@material-ui/icons/VisibilityOff';
-
-import Google from './../../../../assets/images/icons/social-google.svg';
-
-import {strengthColor, strengthIndicator} from '../../../../utils/password-strength';
 
 const useStyles = makeStyles((theme) => ({
     root: {},
@@ -63,7 +58,8 @@ const useStyles = makeStyles((theme) => ({
         marginBottom: theme.spacing(1)
     },
     forgot: {
-        textDecoration: 'none'
+        textDecoration: 'none',
+        color: theme.palette.purple.main
     },
     loginIcon: {
         marginRight: '16px',
@@ -108,19 +104,12 @@ const useStyles = makeStyles((theme) => ({
     }
 }));
 
-const FirebaseRgister = ({className, ...rest}) => {
+const LoginForm = (props, {className, ...rest}) => {
     const classes = useStyles();
-    const scriptedRef = useScriptRef();
     const customization = useSelector((state) => state.customization);
+    const scriptedRef = useScriptRef();
     const [showPassword, setShowPassword] = React.useState(false);
     const [checked, setChecked] = React.useState(true);
-
-    const [strength, setStrength] = React.useState(0);
-    const [level, setLevel] = React.useState('');
-
-    const googleHandler = async () => {
-
-    };
 
     const handleClickShowPassword = () => {
         setShowPassword(!showPassword);
@@ -130,50 +119,13 @@ const FirebaseRgister = ({className, ...rest}) => {
         event.preventDefault();
     };
 
-    const changePassowd = (value) => {
-        const temp = strengthIndicator(value);
-        setStrength(temp);
-        setLevel(strengthColor(temp));
-    };
-
-    useEffect(() => {
-        changePassowd('123456');
-    }, []);
-
     return (
         <React.Fragment>
             <Grid container direction="column" justifyContent="center" spacing={2}>
-                <Grid item xs={12}>
-                    <Button
-                        disableElevation
-                        fullWidth={true}
-                        className={classes.redButton}
-                        onClick={googleHandler}
-                        size="large"
-                        variant="contained"
-                    >
-                        <img src={Google} alt="google" width="20px" className={classes.loginIcon} /> Sign up with Google
-                    </Button>
-                </Grid>
-                <Grid item xs={12}>
-                    <Box alignItems="center" display="flex">
-                        <Divider className={classes.signDivider} orientation="horizontal" />
-                        <Button
-                            variant="outlined"
-                            className={classes.signText}
-                            sx={{borderRadius: customization.borderRadius + 'px'}}
-                            disableRipple
-                            disabled
-                        >
-                            OR
-                        </Button>
-                        <Divider className={classes.signDivider} orientation="horizontal" />
-                    </Box>
-                </Grid>
                 <Grid item xs={12} container alignItems="center" justifyContent="center">
                     <Box mb={2}>
                         <Typography variant="subtitle1" className={classes.title}>
-                            Sign up with Email address
+                            Sign in with Email address
                         </Typography>
                     </Box>
                 </Grid>
@@ -208,47 +160,21 @@ const FirebaseRgister = ({className, ...rest}) => {
             >
                 {({errors, handleBlur, handleChange, handleSubmit, isSubmitting, touched, values}) => (
                     <form noValidate onSubmit={handleSubmit} className={clsx(classes.root, className)} {...rest}>
-                        <Grid container spacing={2}>
-                            <Grid item xs={12} sm={6}>
-                                <TextField
-                                    fullWidth
-                                    label="First Name"
-                                    margin="normal"
-                                    name="fname"
-                                    type="text"
-                                    defaultValue="Joseph"
-                                    variant="outlined"
-                                    className={classes.loginput}
-                                />
-                            </Grid>
-                            <Grid item xs={12} sm={6}>
-                                <TextField
-                                    fullWidth
-                                    label="Last Name"
-                                    margin="normal"
-                                    name="lname"
-                                    type="text"
-                                    defaultValue="Doe"
-                                    variant="outlined"
-                                    className={classes.loginput}
-                                />
-                            </Grid>
-                        </Grid>
                         <FormControl
                             fullWidth
                             error={Boolean(touched.email && errors.email)}
                             className={classes.loginput}
                             variant="outlined"
                         >
-                            <InputLabel htmlFor="outlined-adornment-email-register">Email Address / Username</InputLabel>
+                            <InputLabel htmlFor="outlined-adornment-email-login">Email Address</InputLabel>
                             <OutlinedInput
-                                id="outlined-adornment-email-register"
+                                id="outlined-adornment-email-login"
                                 type="email"
                                 value={values.email}
                                 name="email"
                                 onBlur={handleBlur}
                                 onChange={handleChange}
-                                labelWidth={70}
+                                label='Email Address'
                                 inputProps={{
                                     classes: {
                                         notchedOutline: classes.notchedOutline
@@ -256,7 +182,7 @@ const FirebaseRgister = ({className, ...rest}) => {
                                 }}
                             />
                             {touched.email && errors.email && (
-                                <FormHelperText error id="standard-weight-helper-text--register">
+                                <FormHelperText error id="standard-weight-helper-text-email-login">
                                     {' '}
                                     {errors.email}{' '}
                                 </FormHelperText>
@@ -269,17 +195,14 @@ const FirebaseRgister = ({className, ...rest}) => {
                             className={classes.loginput}
                             variant="outlined"
                         >
-                            <InputLabel htmlFor="outlined-adornment-password-register">Password</InputLabel>
+                            <InputLabel htmlFor="outlined-adornment-password-login">Password</InputLabel>
                             <OutlinedInput
-                                id="outlined-adornment-password-register"
+                                id="outlined-adornment-password-login"
                                 type={showPassword ? 'text' : 'password'}
                                 value={values.password}
                                 name="password"
                                 onBlur={handleBlur}
-                                onChange={(e) => {
-                                    handleChange(e);
-                                    changePassowd(e.target.value);
-                                }}
+                                onChange={handleChange}
                                 endAdornment={
                                     <InputAdornment position="end">
                                         <IconButton
@@ -292,7 +215,7 @@ const FirebaseRgister = ({className, ...rest}) => {
                                         </IconButton>
                                     </InputAdornment>
                                 }
-                                labelWidth={70}
+                                label='Password'
                                 inputProps={{
                                     classes: {
                                         notchedOutline: classes.notchedOutline
@@ -300,30 +223,12 @@ const FirebaseRgister = ({className, ...rest}) => {
                                 }}
                             />
                             {touched.password && errors.password && (
-                                <FormHelperText error id="standard-weight-helper-text-password-register">
+                                <FormHelperText error id="standard-weight-helper-text-password-login">
                                     {' '}
                                     {errors.password}{' '}
                                 </FormHelperText>
                             )}
                         </FormControl>
-
-                        {strength !== 0 && (
-                            <FormControl fullWidth>
-                                <Box mb={2}>
-                                    <Grid container spacing={2} alignItems="center">
-                                        <Grid item>
-                                            <Box width={85} height={8} borderRadius={7} backgroundColor={level.color}></Box>
-                                        </Grid>
-                                        <Grid item>
-                                            <Typography variant="subtitle1" fontSize="0.75rem">
-                                                {level.label}
-                                            </Typography>
-                                        </Grid>
-                                    </Grid>
-                                </Box>
-                            </FormControl>
-                        )}
-
                         <Grid container alignItems="center" justifyContent="space-between">
                             <Grid item>
                                 <FormControlLabel
@@ -335,17 +240,14 @@ const FirebaseRgister = ({className, ...rest}) => {
                                             color="primary"
                                         />
                                     }
-                                    label={
-                                        <React.Fragment>
-                                            <Typography variant="subtitle1">
-                                                Agree with &nbsp;
-                                                <Typography variant="subtitle1" component={Link} to="#">
-                                                    Terms & Condition.
-                                                </Typography>
-                                            </Typography>
-                                        </React.Fragment>
-                                    }
+                                    label={<React.Fragment>Keep me logged in</React.Fragment>}
                                 />
+                            </Grid>
+                            <Grid item>
+                                <Typography
+                                    variant="subtitle1">
+                                    Forgot Password?
+                                </Typography>
                             </Grid>
                         </Grid>
                         {errors.submit && (
@@ -364,7 +266,7 @@ const FirebaseRgister = ({className, ...rest}) => {
                                 variant="contained"
                                 className={classes.login}
                             >
-                                Sign up
+                                Sign in
                             </Button>
                         </Box>
                     </form>
@@ -374,4 +276,4 @@ const FirebaseRgister = ({className, ...rest}) => {
     );
 };
 
-export default FirebaseRgister;
+export default LoginForm;
