--- frontend/src/src/views/sample-page/index.js
@@ -1,26 +1,0 @@
-import React from 'react';
-import {Card, CardContent, CardHeader, Divider, Grid, Typography} from '@material-ui/core';
-
-const SamplePage = () => {
-    return (
-        <Grid container>
-            <Grid item xs={12}>
-                <Card>
-                    <CardHeader title={<Typography variant="h5">Sample Card</Typography>} />
-                    <Divider />
-                    <CardContent>
-                        <Typography variant="body2">
-                            Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore
-                            magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo
-                            consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla
-                            pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est
-                            laborum.
-                        </Typography>
-                    </CardContent>
-                </Card>
-            </Grid>
-        </Grid>
-    );
-};
-
-export default SamplePage;
